Tests work
some changes to make test work

+++ src/main/java/isteriagroup/cryptotracker/dtos/SubscriptionDto.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import isteriagroup.cryptotracker.entities.SubscriptionPK;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @NoArgsConstructor
@@ -17,13 +19,13 @@ public class SubscriptionDto {
 
     @Getter
     @Setter
-    private Float userVal;
+    private BigDecimal userVal;
 
     @Getter
     @Setter
     private CurrencyDto currencyDto;
 
-    public SubscriptionDto(SubscriptionPK subscriptionPK, Float userVal) {
+    public SubscriptionDto(SubscriptionPK subscriptionPK, BigDecimal userVal) {
         this.subscriptionPK = subscriptionPK;
         this.userVal = userVal;
     }
